moved everything to src folder, set up wsgi server (it works)
