+++ src/app.py
@@ -11,12 +11,12 @@ from flask_migrate import Migrate
 from sqlalchemy import exc
 
 from flask_cors import CORS
-from project.models import db, User, Show, Entry
+from .models import db, User, Show, Entry
 
-app = Flask(__name__)
+app = Flask(__name__, template_folder="templates")
 load_dotenv()
 
-app.config.from_pyfile('project/instance/config.py')
+app.config.from_pyfile('instance/config.py')
 # initialize the app with the extension
 app.json.compact = False
